oci-register-machine: Get hook stage from stdin's 'status' value
Instead of using CRI-O-specific arguments or environment variables,
use the state's REQUIRED 'status' property [1] to ensure we're being
called on a created container (and not a running container or a
stopped container).

I've shifted the order in the "Register machine" log entry as well,
because the container ID seems more important than the stage/status.

I've also shifted the "unsupported status" warning outside of the
switch; it had been inside the poststop branch since it landed in
3e9549af7 (Make work with latest hooks patch, 2015-10-29, #2).  Go
switches do not require braces around the per-case blocks [2], but
that may have been the source of the misplacement.

[1]: https://github.com/opencontainers/runtime-spec/blame/v1.0.1/runtime.md#L16
[2]: https://golang.org/ref/spec#Switch_statements